Moved UI core parts of module to ui folder
