+++ core/api/cart/index.js
@@ -1,9 +1,9 @@
 import { addToCart } from './addToCart'
 import { removeFromCart } from './removeFromCart'
 import { productsInCart } from './productsInCart'
-import { closeMicrocart } from './closeMicrocart'
-import { openMicrocart } from './openMicrocart'
-import { isMicrocartOpen } from './isMicrocartOpen'
+import { closeMicrocart } from './ui/closeMicrocart'
+import { openMicrocart } from './ui/openMicrocart'
+import { isMicrocartOpen } from './ui/isMicrocartOpen'
 
 export {
   addToCart,
